requirements.txt
selenium>=4.15.0
webdriver-manager>=4.0.0
beautifulsoup4>=4.12.0
pandas>=2.1.0
streamlit>=1.30.0
matplotlib>=3.8.0
Pillow>=10.0.0
lxml>=4.9.0
numpy>=1.26.0
supabase>=2.0.0
python-dotenv>=1.0.0
